Organize Component and Container file hierarchy
