+++ src/components/Scroll.js
@@ -5,8 +5,7 @@ const Scroll = (props) => {
     <div
       style={{
         overflowY: "scroll",
-        border: "1px solid black",
-        height: "500px",
+        height: "700px",
       }}
     >
       {props.children}
