+++ src/containers/App.css
@@ -4,7 +4,7 @@
     font-family: 'SEGA LOGO FONT';
     font-style: normal;
     font-weight: normal;
-    src: local('SEGA LOGO FONT'), url('SEGA.woff') format('woff');
+    src: local('SEGA LOGO FONT'), url('../components/SEGA.woff') format('woff');
     }
