+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
 import "tachyons";
-import App from "./App";
+import App from "./containers/App";
 
 ReactDOM.render(<App />, document.getElementById("root"));
